Drop `Expense` component
